Added history fragment to display history; retrofit code for it works but logic needs work. TODO: finish logic, PIN confirmation, push out the beta

+++ app/src/main/java/bachmanity/prank_call/API/Models/UserId.java
@@ -9,9 +9,9 @@ import com.google.gson.annotations.SerializedName;
 public class UserId {
     @SerializedName("id")
     @Expose
-    private long userId;
+    private int userId;
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 }
